Move js to be internal, change out line to borders
The defer attribute seemed to be stopping my page
from running, but my nodes were not getting defined
in time without it (at least that is my current assement
of what was happening). Moving it interal allowed me to move
forward with the project to deal with moving the js to
be external later.